LayoutTests/imported/w3c/web-platform-tests/resource-timing/buffer-full-then-increased-expected.txt


FAIL Test that overflowing the buffer and immediately increasing its limit does not trigger the resourcetimingbufferfull event assert_equals: All resources should be in the buffer, since its size was increased expected 3 but got 2
